move config load/save to ServerConfig
