+++ ProjectFiles/Assets/Scripts/Game/ServerGameLoop.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.IO;
-using Network;
-using UnityEngine;
+﻿using Network;
 
 namespace Game
 {
@@ -13,7 +10,7 @@ namespace Game
         public bool Init(string[] args)
         {
             // TODO: make the config path an argument so we can swap configs later?
-            var config = LoadConfigOrDefault("server-config.json");
+            var config = ServerConfig.LoadConfigOrDefault("server-config.json");
             
             // Create world
             world = new World();
@@ -25,34 +22,6 @@ namespace Game
             return success;
         }
 
-        private ServerConfig LoadConfigOrDefault(string path)
-        {
-            try
-            {
-                using (var reader = new StreamReader(path))
-                {
-                    var json = reader.ReadToEnd();
-                    return JsonUtility.FromJson<ServerConfig>(json);
-                }
-            }
-            catch (Exception e) when (e is ArgumentException || e is FileNotFoundException)
-            {
-                Debug.Log(e);
-                Debug.Log($"Failed to load {path}. Generating and running default config at server-config.json...");
-                SaveConfig(path, ServerConfig.DefaultConfig);
-                return ServerConfig.DefaultConfig;
-            }
-        }
-
-        private void SaveConfig(string path, ServerConfig config)
-        {
-            using (var writer = new StreamWriter(path))
-            {
-                var json = JsonUtility.ToJson(config);
-                writer.Write(json);
-            }
-        }
-        
         public void Shutdown()
         {
             server.Shutdown();
